Add utility classes for uppercase and italic
I found myself needing these two frequently.

As a dev, I want to be able to use a utility class to change the style of text without having to make a separate component for it.  I want this because it would save me time, especially for those 'one-off' moments. 

I believe it's low-level enough to become a utility class and it still has a separation from the semantics of a document.